add watch command
A common use case for 'thyme' is to record the active windows at regular intervals.
To make this easier, the command 'watch' can now be used instead of 'track'.

$ thyme watch -n <value> -o thyme.json

should be equivalent to

$ watch -n <value> thyme track -o thyme.json
or
$ while true; do thyme track -o thyme.json; sleep <value>s; done;

but is also usable on systems that do not have 'watch' installed and easier
to use than a shell loop.

If no interval is specified (-n or --interval) then it defaults to 30 seconds.

This was originally suggested by @keegancsmith.